requirements.txt
asgiref==3.5.0
black==22.3.0
click==8.1.2
Django==4.0.4
django-dynamic-fixture==3.1.2
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.5.0
importlib-metadata==4.11.3
Markdown==3.3.6
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
PyJWT==1.7.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
zipp==3.8.0
